feat: move current impl to "single" entry
