+++ .gitignore
@@ -8,3 +8,4 @@ node_modules
 /dist
 /secure
 /index.d.ts
+/single
