+++ src/single.d.ts
