+++ src/single.js
